# api/openapi.yaml
openapi: 3.1.0
info:
  title: portal API
  version: 1.0.0
  description: |
    The portal HTTP API. Every route chi serves under /api/v1
    (internal/server/server.go) is described here; the schemas mirror the
    handler-layer *Response types in internal/handler (the wire boundary —
    repository rows never serialize directly).

    web/src/api/types.ts is generated from this file with openapi-typescript
    (`npm run generate:api` in web/); openapi-fetch consumes the generated
    `paths` interface. Edit this spec, regenerate, and the frontend contract
    follows — CI fails on drift.

    Two surfaces sit outside the JSON request/response model:
    - GET /workspaces/{workspaceId}/runs/{runId}/logs/ws upgrades to a
      WebSocket (documented on the operation; the stream itself is
      newline-delimited log text, authenticated via the
      `bearer` + token subprotocol pair).
    - GET /auth/github and /auth/github/callback are browser redirect hops of
      the OAuth flow, not XHR endpoints.

    Errors use one envelope everywhere (ErrorResponse) via respond.Error /
    respond.FromError; operations declare it as their default response.
servers:
  - url: /api/v1
paths:
...
  /variables:
    get:
      operationId: listOrgVariables
      summary: List org variables
      responses:
        "200":
          description: Org variables (sensitive values redacted)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrgVariableList" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createOrgVariable
      summary: Create an org variable (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVariableRequest" }
      responses:
        "201":
          description: Created variable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrgVariable" }
        default: { $ref: "#/components/responses/Error" }
...
  /pipelines/{pipelineId}/variables:
    get:
      operationId: listPipelineVariables
      summary: List pipeline variables
      parameters:
        - { name: pipelineId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Pipeline variables (sensitive values redacted)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineVariableList" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createPipelineVariable
      summary: Create a pipeline variable (admin)
      parameters:
        - { name: pipelineId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVariableRequest" }
      responses:
        "201":
          description: Created variable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineVariable" }
        default: { $ref: "#/components/responses/Error" }
...
  /workspaces/{workspaceId}/variables:
    get:
      operationId: listWorkspaceVariables
      summary: List workspace variables
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Variables (sensitive values redacted)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceVariableList" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createWorkspaceVariable
      summary: Create a workspace variable (admin)
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVariableRequest" }
      responses:
        "201":
          description: Created variable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceVariable" }
        default: { $ref: "#/components/responses/Error" }
...
  /workspaces/{workspaceId}/variables/{variableId}:
    put:
      operationId: updateWorkspaceVariable
      summary: Update a workspace variable (admin)
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string } }
        - { name: variableId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateVariableRequest" }
      responses:
        "200":
          description: Updated variable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceVariable" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteWorkspaceVariable
      summary: Delete a workspace variable (admin)
      parameters:
        - { name: workspaceId, in: path, required: true, schema: { type: string } }
        - { name: variableId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: The standard error envelope (respond.Error / respond.FromError).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string, description: "HTTP status text." }
        message: { type: string }
        request_id: { type: string, description: "Set on 5xx responses so logs can be correlated." }
      required: [error]
...
    VariableCategory:
      type: string
      enum: [terraform, env]
    WorkspaceVariable:
      type: object
      properties:
        id: { type: string }
        workspace_id: { type: string }
        org_id: { type: string }
        key: { type: string }
        value: { type: string }
        sensitive: { type: boolean }
        category: { $ref: "#/components/schemas/VariableCategory" }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, workspace_id, org_id, key, value, sensitive, category, description, created_at, updated_at]
    OrgVariable:
      type: object
      properties:
        id: { type: string }
        org_id: { type: string }
        key: { type: string }
        value: { type: string }
        sensitive: { type: boolean }
        category: { $ref: "#/components/schemas/VariableCategory" }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, org_id, key, value, sensitive, category, description, created_at, updated_at]
    PipelineVariable:
      type: object
      properties:
        id: { type: string }
        pipeline_id: { type: string }
        org_id: { type: string }
        key: { type: string }
        value: { type: string }
        sensitive: { type: boolean }
        category: { $ref: "#/components/schemas/VariableCategory" }
        description: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, pipeline_id, org_id, key, value, sensitive, category, description, created_at, updated_at]
    CreateVariableRequest:
      type: object
      properties:
        key: { type: string }
        value: { type: string }
        sensitive: { type: boolean }
        category: { $ref: "#/components/schemas/VariableCategory" }
        description: { type: string }
      required: [key, value, sensitive, category]
...
    OrgVariableList:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/OrgVariable" }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
      required: [data, total, page, per_page]
    PipelineVariableList:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/PipelineVariable" }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
      required: [data, total, page, per_page]
    WorkspaceVariableList:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/WorkspaceVariable" }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
      required: [data, total, page, per_page]
...
security:
  - bearerAuth: []